CR changes: removed pandora-test subproject and moved test interface to panodra.core
