+++ gradle/common_deps.gradle
@@ -9,4 +9,5 @@ dependencies {
 
     testCompile "junit:junit:$junitVersion"
     testCompile "org.mockito:mockito-core:$mockitoVersion"
+    testCompile project(':pandora-core')
 }
