+++ pandora-core/src/main/java/com/wikia/pandora/core/test/IntegrationTest.java
@@ -1,4 +1,4 @@
-package com.wikia.pandora.test;
+package com.wikia.pandora.core.test;
 
 public interface IntegrationTest {
 }
