--- pandora-test/build.gradle
@@ -1,1 +1,0 @@
-apply from: rootProject.file('gradle/pandora_common_deps.gradle')
