+++ settings.gradle
@@ -9,5 +9,3 @@ include 'pandora-service'
 include 'mercury-api'
 include 'nirvana-api'
 include 'pandora-example'
-include 'pandora-test'
-
